server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai==0.3.2
sqlalchemy==2.0.27